requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
configparser==5.3.0
confluent-kafka==2.0.2
docopt==0.6.2
idna==3.4
kafka==1.3.5
kafka-python==2.0.1
numpy==1.24.1
pipreqs==0.4.11
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
yarg==0.1.9
opencv-python~=4.7.0.68